правки app, useAuth, useMovie
